latest change

+++ invoices/Models/TaxAndDiscount.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public decimal value { get; set; }
+        public Double value { get; set; }
         public string Type { get; set; }
         public bool IsActive { get; set; }
         public DateTime StartDate { get; set; }
